requirements.txt
openai==0.27.10
Django==4.2.29
djangorestframework==3.15.2
dj_database_url==1.2.0
python-dotenv==1.0.1
whitenoise==6.6.0
django-cors-headers==4.3.0
gunicorn
pydub==0.25.1